+++ src/main/java/kr/fiveminutesmarket/product/error/exception/ProductOptionItemNameDuplicatedException.java
@@ -1,4 +1,4 @@
-package kr.fiveminutesmarket.option.error.exception;
+package kr.fiveminutesmarket.product.error.exception;
 
 public class ProductOptionItemNameDuplicatedException extends RuntimeException{
